move DefPathBasedNames to ty::print::obsolete
