+++ src/librustc/ty/print/mod.rs
@@ -7,6 +7,7 @@ use rustc_data_structures::fx::FxHashSet;
 
 // `pretty` is a separate module only for organization.
 mod pretty;
+pub mod obsolete;
 pub use self::pretty::*;
 
 pub trait Print<'gcx, 'tcx, P> {
